+++ src/client/components/MyTasks/DataHandler.react.js
@@ -16,8 +16,17 @@ import {
 } from './data/fetchers';
 import _ from 'lodash';
 
-export default function withDataHandler(WrappedComponent) {
-
+/**
+ * Injects common invoice operations to wrapped component
+ *
+ * @param WrappedComponent - wrapped component
+ * @param fetcher - list of tasks fetcher
+ * @param filter - predicate defining if invoice should be displayed
+ * @returns {DataHandler}
+ */
+export default function withDataHandler(WrappedComponent,
+                                        { fetcher = () => fetchApprovalTasks({}),
+                                          filter = invoice => !!invoice }) {
   class DataHandler extends Component {
 
     static propTypes = {
@@ -56,7 +65,7 @@ export default function withDataHandler(WrappedComponent) {
 
     componentDidMount() {
       this.loadMasterData().then((masterData) => Promise.resolve(this.setState(masterData, () => {
-        fetchApprovalTasks({ searchParams: { assignedToMe: this.props.assignedToMe } }).then((invoices) => {
+        fetcher().then((invoices) => {
           return Promise.resolve(invoices.length > 0 && this.loadInvoiceData(invoices[0].key).then((invoiceData) => {
             return Promise.resolve(this.setState({ taskList: invoices, invoice: invoiceData }));
           }));
@@ -89,8 +98,8 @@ export default function withDataHandler(WrappedComponent) {
     }
 
     getInvoice(id) {
-      return this.loadInvoiceData(id).then((invoice) => Promise.resolve(this.setState({ invoice }))
-      ).catch((err) => { throw Error(err); });
+      return this.loadInvoiceData(id).then((invoice) =>
+        this.setState({ invoice })).catch((err) => { throw Error(err); });
     }
 
     /**
@@ -112,11 +121,11 @@ export default function withDataHandler(WrappedComponent) {
         }).then(({ invoice, invoiceData }) => {
           let updatedInvoice;
           let updatedTaskList;
-          if (invoiceData.transitions.length === 0 && this.props.assignedToMe) {
-            updatedTaskList = _.filter(this.state.taskList, task => task.key !== invoiceData.key);
-          } else {
+          if (filter(invoiceData)) {
             updatedInvoice = invoiceData;
             updatedTaskList = _.map(this.state.taskList, task => task.key === id ? invoice : task);
+          } else {
+            updatedTaskList = _.filter(this.state.taskList, task => task.key !== invoiceData.key);
           }
           return Promise.resolve(
             this.setState({
