+++ src/client/components/MyTasks/list/MyTasksList.react.js
@@ -22,9 +22,7 @@ export default class MyTasksList extends PureComponent {
 
   componentWillReceiveProps(nextProps) {
     if (this.props.list.length !== nextProps.list.length && nextProps.list.length !== 0) {
-      // Jump one task backward
-      this.setState({ selected: this.state.selected ? this.state.selected - 1 : 0 },
-        () => this.props.getInvoice(nextProps.list[this.state.selected]['key']));
+      this.setState({ selected: 0 }, () => this.props.getInvoice(nextProps.list[0]['key']));
     }
   }
